변경내용 커밋

+++ Source/FlowerParticleShader.h
@@ -6,15 +6,15 @@ class CFlowerParticleShader : public CQuadParticleShader
 {
 private:
 	float m_fPauseTime;
-	int m_fMaxPauseTime;
+	float m_fMaxPauseTime;
 public:
 	CFlowerParticleShader();
 
 	virtual ~CFlowerParticleShader();
 
-	virtual void Initialize(ID3D11Device *pd3dDevice, TCHAR *pszFileName, XMFLOAT3 d3dxvPosition, UINT nMaxParticles);
+	virtual void Initialize(ID3D11Device* pd3dDevice, TCHAR *pszFileName, XMFLOAT3 d3dxvPosition, UINT nMaxParticles);
 
-	virtual void PauseTime(const int& fPause) { m_fMaxPauseTime = fPause; }
+	virtual void PauseTime(const float& fPause) { m_fMaxPauseTime = fPause; }
 
 	virtual bool Update(float fTimeElapsed);
